Refactor: 상품 옵션 로직 리팩터링 상품 옵션 로직 훅으로 분리 Resolves: #16
